backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
llama-index==0.9.14.post3
deeplake==3.8.8
openai==1.3.8
cohere==4.37
httpx<0.28
python-dotenv==1.0.0
cors==1.0.1